requirements.txt
torch
clip-interrogator==0.6.0
unidecode
loguru
strenum
pillow
python-dotenv
pydantic
